+++ lib/types.dart
@@ -1,4 +1,4 @@
-/// This type is used by [AVMediaPlayer], for showing current buffer status
+/// This type is used by [AVMediaPlayer], for showing current buffer status.
 class BufferRange {
   static const empty = BufferRange(0, 0);
 
@@ -7,7 +7,7 @@ class BufferRange {
   const BufferRange(this.begin, this.end);
 }
 
-/// This type is used by [AVMediaPlayer], for showing current media info
+/// This type is used by [AVMediaPlayer], for showing current media info.
 class MediaInfo {
   final int width;
   final int height;
@@ -16,5 +16,5 @@ class MediaInfo {
   const MediaInfo(this.width, this.height, this.duration, this.source);
 }
 
-/// This type is used by [AVMediaPlayer], for showing current playback state
+/// This type is used by [AVMediaPlayer], for showing current playback state.
 enum PlaybackState { playing, paused, closed }