+++ lib/widget.dart
@@ -4,20 +4,25 @@ import 'player.dart';
 /// The widget to display video for [AVMediaPlayer].
 class AVMediaView extends StatefulWidget {
   final AVMediaPlayer? initPlayer;
+  final void Function(AVMediaPlayer player)? onCreated;
+  final Color? backgroundColor;
+  final bool keepAspectRatio;
   final String? initSource;
   final bool? initAutoPlay;
   final bool? initLooping;
   final double? initVolume;
   final double? initSpeed;
   final int? initPosition;
-  final void Function(AVMediaPlayer player)? onCreated;
 
   /// Create a new [AVMediaView] widget.
-  ///
   /// If [initPlayer] is null, a new player will be created. You can get the player from [onCreated] callback.
   ///
-  /// Other parameters are optional. They take efferts when the widget is created.
-  /// And can be changed later by calling the corresponding methods of the player
+  /// [backgroundColor] is the color to display when there is no video. This parameter can be changed by updating the widget.
+  ///
+  /// [keepAspectRatio] is whether to keep the aspect ratio of the video. this parameter can be changed by updating the widget.
+  ///
+  /// Other parameters only take efferts at the time this widget is created.
+  /// To changed them later, you need to calling the corresponding methods of the player.
   const AVMediaView({
     super.key,
     this.initPlayer,
@@ -28,6 +33,8 @@ class AVMediaView extends StatefulWidget {
     this.initSpeed,
     this.initPosition,
     this.onCreated,
+    this.backgroundColor,
+    this.keepAspectRatio = true,
   });
 
   @override
@@ -36,6 +43,7 @@ class AVMediaView extends StatefulWidget {
 
 class _AVMediaState extends State<AVMediaView> {
   bool _foreignPlayer = false;
+  bool _hasVideo = false;
   AVMediaPlayer? _player;
 
   @override
@@ -51,8 +59,9 @@ class _AVMediaState extends State<AVMediaView> {
         initPosition: widget.initPosition,
       );
     } else {
-      _foreignPlayer = true;
       _player = widget.initPlayer!;
+      _foreignPlayer = true;
+      _hasVideo = _checkVideo();
       if (widget.initSource != null) {
         _player!.open(widget.initSource!);
       }
@@ -79,8 +88,15 @@ class _AVMediaState extends State<AVMediaView> {
     if (widget.onCreated != null) {
       widget.onCreated!(_player!);
     }
-    if (_player!.id.value == null) {
-      _player!.id.addListener(_update);
+    _player!.mediaInfo.addListener(_update);
+  }
+
+  @override
+  void didUpdateWidget(AVMediaView oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if ((widget.keepAspectRatio != oldWidget.keepAspectRatio && _hasVideo) ||
+        (widget.backgroundColor != oldWidget.backgroundColor && !_hasVideo)) {
+      setState(() {});
     }
   }
 
@@ -88,11 +104,10 @@ class _AVMediaState extends State<AVMediaView> {
   void dispose() {
     if (_foreignPlayer) {
       try {
+        _player?.mediaInfo.removeListener(_update);
         //maybe the player will be reused by the user.
         //but at least it should be closed to prevent texture link error if there is an open video.
-        if (_player?.mediaInfo.value != null &&
-            _player!.mediaInfo.value!.width > 0 &&
-            _player!.mediaInfo.value!.height > 0) {
+        if (_hasVideo) {
           _player!.close();
         }
       } catch (_) {}
@@ -103,12 +118,20 @@ class _AVMediaState extends State<AVMediaView> {
   }
 
   @override
-  Widget build(BuildContext context) => Container(
-        color: const Color(0xFF000000),
-        child: _player?.id.value == null
-            ? null
-            : Texture(textureId: _player!.id.value!),
-      );
+  Widget build(BuildContext context) {
+    if (_hasVideo) {
+      final texture = Texture(textureId: _player!.id.value!);
+      return widget.keepAspectRatio
+          ? AspectRatio(
+              aspectRatio: _player!.mediaInfo.value!.width /
+                  _player!.mediaInfo.value!.height,
+              child: texture,
+            )
+          : texture;
+    } else {
+      return Container(color: widget.backgroundColor);
+    }
+  }
 
   void _initPosition() {
     _player?.mediaInfo.removeListener(_initPosition);
@@ -117,8 +140,18 @@ class _AVMediaState extends State<AVMediaView> {
     }
   }
 
+  bool _checkVideo() =>
+      _player?.mediaInfo.value != null &&
+      _player!.mediaInfo.value!.width > 0 &&
+      _player!.mediaInfo.value!.height > 0 &&
+      _player!.mediaInfo.value!.duration > 0;
+
   void _update() {
-    _player?.id.removeListener(_update);
-    setState(() {});
+    final hasVideo = _checkVideo();
+    if (_hasVideo != hasVideo) {
+      setState(() {
+        _hasVideo = hasVideo;
+      });
+    }
   }
 }