+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 
 - add dartdoc comments
 - replace default example to the simpe one
+
+## 0.5.3
+
+- imporve dartdoc comments
+- imporve examples
+- combine macos and ios backend code
+- add `backgroundColor` and `keepAspectRatio` support for `AVMediaView` widget